// sygec-frontend/starter/src/app/models/beneficiaire.ts
export class Beneficiaire {
    uuid: string = "";
    service_uuid: string  = "";
    nom: string  = "";
    prenom: string  = "";
    photo: string  = "";
    grade: string  = "";
    email: string  = "";
    phone: string  = "";
    libelleBeneficiaire: string  = "";
    role:any
    uuidimgArcticle: string  = "";
    imgArcticle: string  = "";
    imageUrl: string  = "";
    matricule: string  = "";
}